+++ layers/GatoGraphQLForWP/plugins/gatographql/src/Constants/VirtualRecipes.php
@@ -10,4 +10,5 @@ namespace GatoGraphQL\GatoGraphQL\Constants;
 class VirtualRecipes
 {
     public const IMPORTING_A_POST_FROM_WORDPRESS_RSS_FEED = 'importing-a-post-from-wordpress-rss-feed';
+    public const FETCH_POST_LINKS = 'fetch-post-links';
 }